수동 결제 내역을 수정하는 API 구현

See merge request s09-fintech-finance-sub2/S09P22A108!86

Former-commit-id: fec2c98d16c6b6e47b329f0b37d0e7b2f59dce9c

+++ backend/src/main/java/com/sss/tally/api/calculate/dto/CalculateDto.java
@@ -146,7 +146,7 @@ public class CalculateDto {
 
 		private int myAmount;
 
-		private Double allAmount;
+		private int allAmount;
 
 		private LocalDateTime paymentDate;
